Added Login Ui (#1)
* Initial Commit

This setup is for evan to run the app on an Android device.

* Added Simple Navigation

Now with react-navigation we can move around the app with some button
presses.

* Started Login Menu UI

Make it look pretty, it's the first page after all

* Added Sign In Capability

Looks pretty and changes with input

* Finished Login UI

We now have two views which are login and sign up, where you can
interact with both views. Passwords can be shown and hidden.