Added Expense Editing to EveryReceipt (#19)
* Made Styling Scalable for Application

Big big big commit. I moved a styles.js file and a common component for
a button to a Common directory. This makes it easier to find and use
both the button and styling. Plus, when we want to add more styles, we
just add another entry into the styles.js line.

* Added Style to Item

I added a touchable opacity on the button component. Now we can handle
the touch by bringing the user to a screen to edit the selected item. I
also moved the delete button to the right and made it smaller, so the
user doesn't press it on accident.

* Added Screen for Item Edit

Now, when we click the item's touchable opacity, the user is presented
with a screen. Next move is to populate the screen with the selected
item's information.

* Added Item Edit Conent + Spacing to Expense List

Now the user can see all of the information about their item on the edit
page, and when we add multiple items the an expense list, it is layed
out with some space, preventing overlap.

* Made Forms More Scalable

Now we have a component that can be reused to both add an item and edit
the data.

* Went back a step

Double components is too crazy, lets localize it and just deal with the
issue of asynchronous setState.

* Added Default Values to Text Inputs

Now when the user wishes to update a field for their item, the textboxes
already have the values of the existing item, but as a default value.

* Finished Expense Edit Functionality

Added proper logic in form fields as well as item edit to handle the
change. I also added the edit functionality for our database in expense actions. Final boss battle: search.