Added Search to EveryReceipt (#21)
* Started Item Search

We will go to a seperate view to search the items by store and total
value.

* Made Expense List More Scalable

Now, ExpenseList is in the common file, which simply maps whatever we
send to it. This is so we can reuse the file for showing search results
and all expenses.

* Made Expense List more Scalable

Now we can reuse it for two different screens. Next step is to allow
users to go back to the search screen from ItemEdit.js

* Added Search to EveryReceipt

Now, the user is brought to a page where they can query their expenses.
They can search expenses by store and by price, being either less than
or greater than. The price incorporates amounts that are equal to that
price as well.

* Fixed Delete and Search Bug

When deleting an item in the search results list, the item would remain,
same with if the item was edited