Added Receipt Camera Parsing (#35)
* Started Cloud Vision Text Read

* Added Better Cropping to Image

Now we can crop images to whatever size we want. Parsing text is next.

* Added Ability to Grab Text Response

Now we parse the response to only grab the text that Google saw. On to the next step.

* Added Total Extraction

We now can extract a total price from the OCR scan. Next is moving the
total to FormFields to add stuff to it. Yay

* Added No Value Handling To Camera

If the uri is undefined, we leave and do nothing.

* Added Cloud Vision File

Now you can never see it

* Added Proper Undefined Value Handling

Now we will only display real items in the expense list and alert the
user if we couldn't get anything from a scan.

* Finished Receipt Parsing

We now can take a photo of a receipt and from there get the values. DID
THAT!!!!!!!!!!!!!!!!

* Fixed SetState Bug

Wierd loop inariant, now its ok. Were lamost done

* removed console.log